pyproject.toml
[project]
name = "dspider"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pika>=1.3.2",
    "playwright>=1.56.0",
    "pymongo>=4.15.3",
    "pymysql>=1.1.2",
    "pytest>=9.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scrapy>=2.13.4",
]
